chore(wmg): refactor Filters and child components (#6046)
