+++ frontend/src/views/WheresMyGeneV2/components/Filters/components/ColorScale/style.ts
