+++ frontend/src/views/WheresMyGeneV2/components/Filters/components/Compare/style.ts
