+++ frontend/src/views/WheresMyGeneV2/components/Filters/components/Organism/connect.ts
@@ -1,29 +1,15 @@
-import {
-  DefaultMenuSelectOption,
-  InputDropdownProps as RawInputDropdownProps,
-} from "@czi-sds/components";
 import { useContext, useEffect, useMemo } from "react";
-import { track } from "src/common/analytics";
-import { EVENTS } from "src/common/analytics/events";
-import { EMPTY_ARRAY } from "src/common/constants/utils";
+import { Organism as IOrganism } from "src/views/WheresMyGene/common/types";
 import { useAvailableOrganisms } from "src/common/queries/wheresMyGene";
 import {
   DispatchContext,
   StateContext,
 } from "src/views/WheresMyGene/common/store";
 import { selectOrganism } from "src/views/WheresMyGene/common/store/actions";
-import { Organism as IOrganism } from "src/views/WheresMyGene/common/types";
-import { StyledDropdown, Wrapper, Label } from "../common/style";
+import { track } from "src/common/analytics";
+import { EVENTS } from "src/common/analytics/events";
 
-const InputDropdownProps: Partial<RawInputDropdownProps> = {
-  sdsStyle: "square",
-};
-
-interface Props {
-  isLoading: boolean;
-}
-
-export default function Organism({ isLoading }: Props): JSX.Element {
+export const useConnect = () => {
   const dispatch = useContext(DispatchContext);
   const { selectedOrganismId } = useContext(StateContext);
 
@@ -54,21 +40,6 @@ export default function Organism({ isLoading }: Props): JSX.Element {
     return result;
   }, [organisms]);
 
-  const organism = organismsById[selectedOrganismId || ""];
-  return (
-    <Wrapper>
-      <Label>Organism</Label>
-      <StyledDropdown
-        label={organism?.name || "Select"}
-        options={organisms || EMPTY_ARRAY}
-        onChange={handleOnChange as tempOnChange}
-        InputDropdownProps={{ ...InputDropdownProps, disabled: isLoading }}
-        data-testid="add-organism"
-        value={organism}
-      />
-    </Wrapper>
-  );
-
   function handleOnChange(organism: IOrganism | null): void {
     if (!dispatch || !organism || selectedOrganismId === organism.id) return;
 
@@ -76,9 +47,7 @@ export default function Organism({ isLoading }: Props): JSX.Element {
 
     dispatch(selectOrganism(organism?.id || null));
   }
-}
 
-// (HACK): Not sure why styled Dropdown changes `onChange` type
-type tempOnChange = (
-  options: DefaultMenuSelectOption | DefaultMenuSelectOption[] | null
-) => void;
+  const organism = organismsById[selectedOrganismId || ""];
+  return { organisms, organism, handleOnChange };
+};
