+++ frontend/src/views/WheresMyGeneV2/components/Filters/components/Sort/types.ts
@@ -1,0 +1,3 @@
+export interface Props {
+  areFiltersDisabled: boolean;
+}
